refactor(ide): extract mode-agnostic IdeShell from /ide route
Lift the IDE layout (editor, file tree, terminals, portal, resize, mobile tabs, overlays) and its view/lifecycle state out of routes/ide/+page.svelte into a reusable IdeShell component. The route is now a thin wrapper owning only mode-specific concerns (session creation, ?framework= parsing, navigation), passing a boot closure + onSelectFramework into the shell. Lets a future GitHub-clone route render the same shell with a different boot closure. No behavior change.
